refactor(actions): split worktreeActions.ts by cluster
- Extract 7 cluster files (query, service, create, navigation, context, GitHub, resource)
- Rewrite worktreeActions.ts as delegating barrel preserving the public API
- Zero behavior changes, zero changes to importers or tests
