+++ src/services/actions/definitions/worktreeActions.ts
@@ -1,1172 +1,18 @@
 import type { ActionCallbacks, ActionRegistry } from "../actionTypes";
-import { defineAction } from "../defineAction";
-import { z } from "zod";
-import type { ActionContext, ActionId } from "@shared/types/actions";
-import { actionService } from "@/services/ActionService";
-import { copyTreeClient, githubClient, systemClient, worktreeClient } from "@/clients";
-import { getCurrentViewStore, getCurrentViewStoreOrNull } from "@/store/createWorktreeStore";
-import { useWorktreeSelectionStore } from "@/store/worktreeStore";
-import { DEFAULT_COPYTREE_FORMAT } from "@/lib/copyTreeFormat";
-import { notify } from "@/lib/notify";
-import { formatErrorMessage } from "@shared/utils/errorMessage";
-import { getVisibleWorktreesForCycling } from "@/lib/worktreeCyclingOrder";
-import { logError, logWarn } from "@/utils/logger";
-
-function notifyWorktreeResourceError(err: unknown, title: string, fallbackMessage: string): void {
-  const message = formatErrorMessage(err, fallbackMessage) || fallbackMessage;
-  notify({
-    type: "error",
-    priority: "high",
-    title,
-    message,
-    action: {
-      label: "Copy details",
-      successLabel: "Copied",
-      onClick: async () => {
-        try {
-          await navigator.clipboard.writeText(message);
-        } catch {
-          // clipboard write is non-critical
-        }
-      },
-    },
-  });
-}
+import { registerWorktreeQueryActions } from "./worktreeQueryActions";
+import { registerWorktreeServiceActions } from "./worktreeServiceActions";
+import { registerWorktreeCreateActions } from "./worktreeCreateActions";
+import { registerWorktreeNavigationActions } from "./worktreeNavigationActions";
+import { registerWorktreeContextActions } from "./worktreeContextActions";
+import { registerWorktreeGitHubActions } from "./worktreeGitHubActions";
+import { registerWorktreeResourceActions } from "./worktreeResourceActions";
 
 export function registerWorktreeActions(actions: ActionRegistry, callbacks: ActionCallbacks): void {
-  // Query action: list all worktrees with metadata
-  actions.set("worktree.list", () => ({
-    id: "worktree.list",
-    title: "List Worktrees",
-    description: "Get list of all worktrees with status information",
-    category: "worktree",
-    kind: "query",
-    danger: "safe",
-    scope: "renderer",
-    run: async () => {
-      const worktrees = callbacks.getWorktrees();
-      const activeWorktreeId = callbacks.getActiveWorktreeId();
-
-      return worktrees.map((w) => ({
-        id: w.id,
-        path: w.path,
-        branch: w.branch,
-        isActive: w.id === activeWorktreeId,
-        isMain: w.isMainWorktree ?? false,
-        issueNumber: w.issueNumber ?? null,
-        issueTitle: w.issueTitle ?? null,
-        prNumber: w.prNumber ?? null,
-        prTitle: w.prTitle ?? null,
-        prUrl: w.prUrl ?? null,
-        status: w.mood ?? null,
-        lastCommit: w.summary ?? null,
-      }));
-    },
-  }));
-
-  // Query action: get current/active worktree details
-  actions.set("worktree.getCurrent", () => ({
-    id: "worktree.getCurrent",
-    title: "Get Current Worktree",
-    description: "Get the currently active worktree details",
-    category: "worktree",
-    kind: "query",
-    danger: "safe",
-    scope: "renderer",
-    run: async () => {
-      const activeWorktreeId = callbacks.getActiveWorktreeId();
-      if (!activeWorktreeId) {
-        return null;
-      }
-
-      const worktree = getCurrentViewStore().getState().worktrees.get(activeWorktreeId);
-      if (!worktree) {
-        return null;
-      }
-
-      return {
-        id: worktree.id,
-        path: worktree.path,
-        branch: worktree.branch,
-        isMain: worktree.isMainWorktree ?? false,
-        issueNumber: worktree.issueNumber ?? null,
-        issueTitle: worktree.issueTitle ?? null,
-        prNumber: worktree.prNumber ?? null,
-        prTitle: worktree.prTitle ?? null,
-        prUrl: worktree.prUrl ?? null,
-        status: worktree.mood ?? null,
-        lastCommit: worktree.summary ?? null,
-      };
-    },
-  }));
-
-  actions.set("worktree.refresh", () => ({
-    id: "worktree.refresh",
-    title: "Refresh Sidebar",
-    description: "Refresh worktrees, pull requests, and GitHub stats",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    keywords: ["sync", "reload", "update", "sidebar"],
-    run: async () => {
-      window.dispatchEvent(new CustomEvent("daintree:refresh-sidebar"));
-      await Promise.allSettled([
-        window.electron.worktreePort.request("refresh"),
-        worktreeClient.refreshPullRequests(),
-      ]);
-    },
-  }));
-
-  actions.set("worktree.refreshPullRequests", () => ({
-    id: "worktree.refreshPullRequests",
-    title: "Refresh Pull Requests",
-    description: "Refresh PR information for all worktrees",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    keywords: ["pr", "github", "fetch", "sync"],
-    run: async () => {
-      await worktreeClient.refreshPullRequests();
-    },
-  }));
-
-  actions.set("worktree.restartService", () => ({
-    id: "worktree.restartService",
-    title: "Restart Workspace Service",
-    description:
-      "Restart the workspace host. Available after the service has crashed and could not recover automatically.",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    keywords: ["workspace", "backend", "recover", "host"],
-    isEnabled: () => {
-      const store = getCurrentViewStoreOrNull();
-      return store !== null && store.getState().error !== null;
-    },
-    disabledReason: () => {
-      const store = getCurrentViewStoreOrNull();
-      if (store === null) return "No project view available";
-      if (store.getState().error === null) return "Workspace service has not crashed";
-      return undefined;
-    },
-    run: async () => {
-      await worktreeClient.restartService();
-    },
-  }));
-
-  actions.set("worktree.setActive", () =>
-    defineAction({
-      id: "worktree.setActive",
-      title: "Set Active Worktree",
-      description: "Set the active worktree in the backend",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ worktreeId: z.string() }),
-      run: async ({ worktreeId }) => {
-        await worktreeClient.setActive(worktreeId);
-      },
-    })
-  );
-
-  actions.set("worktree.quickCreate", () => ({
-    id: "worktree.quickCreate",
-    title: "Quick Create Worktree",
-    description: "Open recipe picker for quick worktree creation",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    keywords: ["new", "branch", "checkout", "recipe"],
-    run: async () => {
-      useWorktreeSelectionStore.getState().openQuickCreate();
-    },
-  }));
-
-  actions.set("worktree.createDialog.open", () => ({
-    id: "worktree.createDialog.open",
-    title: "New Worktree",
-    description: "Open dialog to create a new worktree",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    keywords: ["create", "branch", "checkout", "add"],
-    run: async () => {
-      useWorktreeSelectionStore.getState().openCreateDialog();
-    },
-  }));
-
-  actions.set("worktree.create", () =>
-    defineAction({
-      id: "worktree.create",
-      title: "Create Worktree",
-      description: "Create a new worktree",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({
-        rootPath: z.string().describe("Root path of the git repository"),
-        options: z
-          .object({
-            baseBranch: z.string().describe("Branch to base the worktree on"),
-            newBranch: z.string().describe("Name for the new branch"),
-            path: z.string().describe("Filesystem path for the new worktree"),
-            fromRemote: z.boolean().optional().describe("Whether baseBranch is a remote branch"),
-            useExistingBranch: z
-              .boolean()
-              .optional()
-              .describe("Use an existing branch instead of creating a new one"),
-            provisionResource: z
-              .boolean()
-              .optional()
-              .describe("Run resource.provision after setup"),
-            worktreeMode: z
-              .string()
-              .optional()
-              .describe('Worktree environment mode ("local" or an environment key)'),
-          })
-          .describe("Worktree creation options"),
-      }),
-      resultSchema: z.string(),
-      run: async ({ rootPath, options }) => {
-        const worktreeId = await worktreeClient.create(options, rootPath);
-        if (!worktreeId) {
-          throw new Error("Failed to create worktree: no worktreeId returned from backend");
-        }
-        return worktreeId;
-      },
-    })
-  );
-
-  actions.set("worktree.listBranches", () =>
-    defineAction({
-      id: "worktree.listBranches",
-      title: "List Branches",
-      description: "List git branches for a repository root",
-      category: "worktree",
-      kind: "query",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ rootPath: z.string() }),
-      run: async ({ rootPath }) => {
-        return await worktreeClient.listBranches(rootPath);
-      },
-    })
-  );
-
-  actions.set("worktree.getDefaultPath", () =>
-    defineAction({
-      id: "worktree.getDefaultPath",
-      title: "Get Default Worktree Path",
-      description: "Get the default path for a new worktree based on branch and config",
-      category: "worktree",
-      kind: "query",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ rootPath: z.string(), branchName: z.string() }),
-      run: async ({ rootPath, branchName }) => {
-        return await worktreeClient.getDefaultPath(rootPath, branchName);
-      },
-    })
-  );
-
-  actions.set("worktree.delete", () =>
-    defineAction({
-      id: "worktree.delete",
-      title: "Delete Worktree",
-      description: "Delete a worktree",
-      category: "worktree",
-      kind: "command",
-      danger: "confirm",
-      scope: "renderer",
-      argsSchema: z.object({
-        worktreeId: z.string(),
-        force: z.boolean().optional(),
-        deleteBranch: z.boolean().optional(),
-      }),
-      run: async ({ worktreeId, force, deleteBranch }) => {
-        await worktreeClient.delete(worktreeId, force, deleteBranch);
-      },
-    })
-  );
-
-  actions.set("worktree.getAvailableBranch", () =>
-    defineAction({
-      id: "worktree.getAvailableBranch",
-      title: "Get Available Branch Name",
-      description:
-        "Get a collision-safe branch name. Returns the original name if available, or a numbered variant if the branch exists.",
-      category: "worktree",
-      kind: "query",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ rootPath: z.string(), branchName: z.string() }),
-      run: async ({ rootPath, branchName }) => {
-        return await worktreeClient.getAvailableBranch(rootPath, branchName);
-      },
-    })
-  );
-
-  actions.set("worktree.select", () =>
-    defineAction({
-      id: "worktree.select",
-      title: "Select Worktree",
-      description: "Select a worktree by ID",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      keywords: ["choose", "activate", "focus", "switch"],
-      argsSchema: z.object({ worktreeId: z.string().optional() }).optional(),
-      nonRepeatable: true,
-      run: async (args, ctx: ActionContext) => {
-        const worktreeId = args?.worktreeId;
-        const targetWorktreeId = worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!targetWorktreeId) {
-          throw new Error("No worktree selected");
-        }
-        useWorktreeSelectionStore.getState().selectWorktree(targetWorktreeId);
-      },
-    })
-  );
-
-  actions.set("worktree.next", () => ({
-    id: "worktree.next",
-    title: "Next Worktree",
-    description: "Switch to the next worktree",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    keywords: ["cycle", "forward", "advance", "switch"],
-    nonRepeatable: true,
-    run: async () => {
-      const activeWorktreeId = callbacks.getActiveWorktreeId();
-      const worktrees = getVisibleWorktreesForCycling(activeWorktreeId);
-      if (worktrees.length === 0) return;
-      const currentIndex = activeWorktreeId
-        ? worktrees.findIndex((w) => w.id === activeWorktreeId)
-        : -1;
-      const nextIndex = currentIndex === -1 ? 0 : (currentIndex + 1) % worktrees.length;
-      useWorktreeSelectionStore.getState().selectWorktree(worktrees[nextIndex]!.id);
-    },
-  }));
-
-  actions.set("worktree.previous", () => ({
-    id: "worktree.previous",
-    title: "Previous Worktree",
-    description: "Switch to the previous worktree",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    keywords: ["cycle", "back", "switch", "last"],
-    nonRepeatable: true,
-    run: async () => {
-      const activeWorktreeId = callbacks.getActiveWorktreeId();
-      const worktrees = getVisibleWorktreesForCycling(activeWorktreeId);
-      if (worktrees.length === 0) return;
-      const currentIndex = activeWorktreeId
-        ? worktrees.findIndex((w) => w.id === activeWorktreeId)
-        : -1;
-      // When the active worktree is outside the visible list, wrap from the
-      // end — this matches worktree.up/upVim so directional and cycle actions
-      // agree when filters hide the active worktree.
-      const prevIndex =
-        currentIndex === -1
-          ? worktrees.length - 1
-          : (currentIndex - 1 + worktrees.length) % worktrees.length;
-      useWorktreeSelectionStore.getState().selectWorktree(worktrees[prevIndex]!.id);
-    },
-  }));
-
-  // Worktree switch by index (parameterized)
-  actions.set("worktree.switchIndex", () =>
-    defineAction({
-      id: "worktree.switchIndex",
-      title: "Switch to Worktree by Index",
-      description: "Switch to worktree at a specific position (1-9)",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ index: z.number().int().min(1).max(9) }),
-      nonRepeatable: true,
-      run: async ({ index }) => {
-        const activeWorktreeId = callbacks.getActiveWorktreeId();
-        const worktrees = getVisibleWorktreesForCycling(activeWorktreeId);
-        if (worktrees.length >= index) {
-          useWorktreeSelectionStore.getState().selectWorktree(worktrees[index - 1]!.id);
-        }
-      },
-    })
-  );
-
-  // Non-parameterized worktree switch actions (for KeyAction/keybinding compatibility)
-  for (let index = 1; index <= 9; index++) {
-    const actionId = `worktree.switch${index}` as ActionId;
-    actions.set(actionId, () => ({
-      id: actionId,
-      title: `Switch to Worktree ${index}`,
-      description: `Switch to worktree at position ${index}`,
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      nonRepeatable: true,
-      run: async () => {
-        const activeWorktreeId = callbacks.getActiveWorktreeId();
-        const worktrees = getVisibleWorktreesForCycling(activeWorktreeId);
-        if (worktrees.length >= index) {
-          useWorktreeSelectionStore.getState().selectWorktree(worktrees[index - 1]!.id);
-        }
-      },
-    }));
-  }
-
-  const selectWorktreeByOffset = (offset: number) => {
-    const activeWorktreeId = callbacks.getActiveWorktreeId();
-    const worktrees = getVisibleWorktreesForCycling(activeWorktreeId);
-    if (worktrees.length === 0) return;
-    const currentIndex = activeWorktreeId
-      ? worktrees.findIndex((w) => w.id === activeWorktreeId)
-      : -1;
-    let nextIndex: number;
-    if (currentIndex === -1) {
-      // Active worktree is outside the visible list (filtered out or unset).
-      // Moving down lands on the first visible entry; moving up lands on the last.
-      nextIndex = offset > 0 ? 0 : worktrees.length - 1;
-    } else {
-      nextIndex = currentIndex + offset;
-      if (nextIndex < 0 || nextIndex >= worktrees.length) return;
-    }
-    useWorktreeSelectionStore.getState().selectWorktree(worktrees[nextIndex]!.id);
-  };
-
-  actions.set("worktree.up", () => ({
-    id: "worktree.up",
-    title: "Worktree Up",
-    description: "Move selection up in the worktree list",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    nonRepeatable: true,
-    run: async () => {
-      selectWorktreeByOffset(-1);
-    },
-  }));
-
-  actions.set("worktree.down", () => ({
-    id: "worktree.down",
-    title: "Worktree Down",
-    description: "Move selection down in the worktree list",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    nonRepeatable: true,
-    run: async () => {
-      selectWorktreeByOffset(1);
-    },
-  }));
-
-  actions.set("worktree.upVim", () => ({
-    id: "worktree.upVim",
-    title: "Worktree Up (Vim)",
-    description: "Move selection up in the worktree list",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    nonRepeatable: true,
-    run: async () => {
-      selectWorktreeByOffset(-1);
-    },
-  }));
-
-  actions.set("worktree.downVim", () => ({
-    id: "worktree.downVim",
-    title: "Worktree Down (Vim)",
-    description: "Move selection down in the worktree list",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    nonRepeatable: true,
-    run: async () => {
-      selectWorktreeByOffset(1);
-    },
-  }));
-
-  actions.set("worktree.home", () => ({
-    id: "worktree.home",
-    title: "Worktree Home",
-    description: "Select the first worktree",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    nonRepeatable: true,
-    run: async () => {
-      const worktrees = getVisibleWorktreesForCycling(callbacks.getActiveWorktreeId());
-      if (worktrees.length === 0) return;
-      useWorktreeSelectionStore.getState().selectWorktree(worktrees[0]!.id);
-    },
-  }));
-
-  actions.set("worktree.end", () => ({
-    id: "worktree.end",
-    title: "Worktree End",
-    description: "Select the last worktree",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    nonRepeatable: true,
-    run: async () => {
-      const worktrees = getVisibleWorktreesForCycling(callbacks.getActiveWorktreeId());
-      if (worktrees.length === 0) return;
-      useWorktreeSelectionStore.getState().selectWorktree(worktrees[worktrees.length - 1]!.id);
-    },
-  }));
-
-  actions.set("worktree.selectSpace", () => ({
-    id: "worktree.selectSpace",
-    title: "Select Worktree (Space)",
-    description: "Select the currently focused worktree",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    nonRepeatable: true,
-    run: async () => {
-      const focused = useWorktreeSelectionStore.getState().focusedWorktreeId;
-      if (!focused) return;
-      useWorktreeSelectionStore.getState().selectWorktree(focused);
-    },
-  }));
-
-  actions.set("worktree.openPalette", () => ({
-    id: "worktree.openPalette",
-    title: "Open Worktree Palette",
-    description: "Open the worktree selection palette",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    keywords: ["switcher", "chooser", "list", "picker"],
-    nonRepeatable: true,
-    run: async () => {
-      callbacks.onOpenWorktreePalette();
-    },
-  }));
-
-  actions.set("worktree.overview", () => ({
-    id: "worktree.overview",
-    title: "Toggle Worktree Overview",
-    description: "Open or close the full-screen worktree overview modal",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    keywords: ["dashboard", "grid", "summary", "modal"],
-    nonRepeatable: true,
-    run: async () => {
-      callbacks.onToggleWorktreeOverview();
-    },
-  }));
-
-  actions.set("worktree.overview.open", () => ({
-    id: "worktree.overview.open",
-    title: "Open Worktree Overview",
-    description: "Open the full-screen worktree overview modal",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    keywords: ["dashboard", "grid", "summary", "modal"],
-    nonRepeatable: true,
-    run: async () => {
-      callbacks.onOpenWorktreeOverview();
-    },
-  }));
-
-  actions.set("worktree.overview.close", () => ({
-    id: "worktree.overview.close",
-    title: "Close Worktree Overview",
-    description: "Close the full-screen worktree overview modal",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    keywords: ["dismiss", "hide", "exit", "escape"],
-    nonRepeatable: true,
-    run: async () => {
-      callbacks.onCloseWorktreeOverview();
-    },
-  }));
-
-  actions.set("worktree.panel", () => ({
-    id: "worktree.panel",
-    title: "Open Worktree Panel",
-    description: "Open the worktree panel",
-    category: "worktree",
-    kind: "command",
-    danger: "safe",
-    scope: "renderer",
-    keywords: ["sidebar", "switcher", "chooser", "picker"],
-    nonRepeatable: true,
-    run: async () => {
-      callbacks.onOpenWorktreePalette();
-    },
-  }));
-
-  actions.set("worktree.copyTree", () =>
-    defineAction({
-      id: "worktree.copyTree",
-      title: "Copy Worktree Context",
-      description: "Generate and copy context for a worktree to clipboard",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z
-        .object({
-          worktreeId: z.string().optional(),
-          format: z.enum(["xml", "json", "markdown", "tree", "ndjson"]).optional(),
-          modified: z.boolean().optional(),
-        })
-        .optional(),
-      run: async (args, ctx: ActionContext) => {
-        const worktreeId = args?.worktreeId;
-        const explicitFormat = args?.format;
-        const modified = args?.modified;
-        const targetWorktreeId = worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!targetWorktreeId) return null;
-
-        const format = explicitFormat ?? DEFAULT_COPYTREE_FORMAT;
-
-        const result = await copyTreeClient.generateAndCopyFile(targetWorktreeId, {
-          format,
-          modified,
-        });
-
-        if (result.error) {
-          if (modified && result.error.includes("No valid files")) {
-            throw new Error("No modified files to copy. Make some changes first.");
-          }
-          throw new Error(result.error);
-        }
-
-        return {
-          worktreeId: targetWorktreeId,
-          fileCount: result.fileCount,
-          stats: result.stats ?? null,
-          format,
-        };
-      },
-    })
-  );
-
-  actions.set("worktree.copyContext", () =>
-    defineAction({
-      id: "worktree.copyContext",
-      title: "Copy Worktree Context (Alias)",
-      description: "Alias for worktree.copyTree",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z
-        .object({
-          worktreeId: z.string().optional(),
-          format: z.enum(["xml", "json", "markdown", "tree", "ndjson"]).optional(),
-          modified: z.boolean().optional(),
-        })
-        .optional(),
-      run: async (args) => {
-        const result = await actionService.dispatch("worktree.copyTree", args, { source: "user" });
-        if (!result.ok) {
-          throw new Error(result.error.message);
-        }
-        return result.result as unknown;
-      },
-    })
-  );
-
-  actions.set("worktree.inject", () =>
-    defineAction({
-      id: "worktree.inject",
-      title: "Inject Worktree Context into Focused Terminal",
-      description: "Inject this worktree's context into the currently focused terminal",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z
-        .object({
-          worktreeId: z.string().optional(),
-        })
-        .optional(),
-      isEnabled: (ctx: ActionContext) => {
-        const hasFocusedTerminal = Boolean(ctx.focusedTerminalId);
-        return hasFocusedTerminal;
-      },
-      disabledReason: (ctx: ActionContext) => {
-        const hasFocusedTerminal = Boolean(ctx.focusedTerminalId);
-        if (!hasFocusedTerminal) {
-          return "No focused terminal to inject into";
-        }
-        return undefined;
-      },
-      run: async (args, ctx: ActionContext) => {
-        const hasFocusedTerminal = Boolean(ctx.focusedTerminalId);
-        if (!hasFocusedTerminal) {
-          throw new Error("No focused terminal to inject into");
-        }
-        const worktreeId = args?.worktreeId;
-        const targetWorktreeId = worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!targetWorktreeId) {
-          throw new Error("No worktree selected");
-        }
-        callbacks.onInject(targetWorktreeId);
-      },
-    })
-  );
-
-  actions.set("worktree.openEditor", () =>
-    defineAction({
-      id: "worktree.openEditor",
-      title: "Open in Editor",
-      description: "Open a worktree folder in the OS file manager / editor",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ worktreeId: z.string().optional() }).optional(),
-      run: async (args, ctx: ActionContext) => {
-        const worktreeId = args?.worktreeId;
-        const targetWorktreeId = worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!targetWorktreeId) return;
-
-        const worktree = getCurrentViewStore().getState().worktrees.get(targetWorktreeId);
-        if (!worktree) return;
-
-        await systemClient.openPath(worktree.path);
-      },
-    })
-  );
-
-  actions.set("worktree.reveal", () =>
-    defineAction({
-      id: "worktree.reveal",
-      title: "Reveal Worktree",
-      description: "Reveal a worktree folder in the OS file manager",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ worktreeId: z.string().optional() }).optional(),
-      run: async (args, ctx: ActionContext) => {
-        const worktreeId = args?.worktreeId;
-        const targetWorktreeId = worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!targetWorktreeId) return;
-        const worktree = getCurrentViewStore().getState().worktrees.get(targetWorktreeId);
-        if (!worktree) return;
-        await systemClient.openPath(worktree.path);
-      },
-    })
-  );
-
-  actions.set("worktree.openIssue", () =>
-    defineAction({
-      id: "worktree.openIssue",
-      title: "Open Worktree Issue",
-      description: "Open the GitHub issue associated with a worktree",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ worktreeId: z.string().optional() }).optional(),
-      run: async (args, ctx: ActionContext) => {
-        const worktreeId = args?.worktreeId;
-        const targetWorktreeId = worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!targetWorktreeId) return;
-        const worktree = getCurrentViewStore().getState().worktrees.get(targetWorktreeId);
-        if (!worktree?.issueNumber) return;
-        await githubClient.openIssue(worktree.path, worktree.issueNumber);
-      },
-    })
-  );
-
-  actions.set("worktree.openPR", () =>
-    defineAction({
-      id: "worktree.openPR",
-      title: "Open Worktree Pull Request",
-      description: "Open the GitHub pull request associated with a worktree",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ worktreeId: z.string().optional() }).optional(),
-      run: async (args, ctx: ActionContext) => {
-        const worktreeId = args?.worktreeId;
-        const targetWorktreeId = worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!targetWorktreeId) return;
-        const worktree = getCurrentViewStore().getState().worktrees.get(targetWorktreeId);
-        if (!worktree?.prUrl) return;
-        await githubClient.openPR(worktree.prUrl);
-      },
-    })
-  );
-
-  actions.set("worktree.openPRInPortal", () =>
-    defineAction({
-      id: "worktree.openPRInPortal",
-      title: "Open Worktree PR in Portal",
-      description: "Open the worktree's GitHub pull request in the integrated browser",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ worktreeId: z.string().optional() }).optional(),
-      run: async (args, ctx: ActionContext) => {
-        const worktreeId = args?.worktreeId;
-        const targetWorktreeId = worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!targetWorktreeId) return;
-
-        const worktree = getCurrentViewStore().getState().worktrees.get(targetWorktreeId);
-        if (!worktree?.prUrl) return;
-
-        try {
-          const url = new URL(worktree.prUrl);
-          if (!["https:", "http:"].includes(url.protocol)) {
-            logWarn(`Invalid PR URL protocol: ${url.protocol}`);
-            return;
-          }
-        } catch (error) {
-          logError(`Invalid PR URL: ${worktree.prUrl}`, error);
-          return;
-        }
-
-        await actionService.dispatch(
-          "portal.openUrl",
-          {
-            url: worktree.prUrl,
-            title: worktree.prTitle || `PR #${worktree.prNumber}`,
-            background: false,
-          },
-          { source: "user" }
-        );
-      },
-      isEnabled: (ctx: ActionContext) => {
-        const worktreeId = ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!worktreeId) return false;
-        const worktree = getCurrentViewStore().getState().worktrees.get(worktreeId);
-        return typeof worktree?.prUrl === "string" && worktree.prUrl.trim().length > 0;
-      },
-      disabledReason: (ctx: ActionContext) => {
-        const worktreeId = ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!worktreeId) return "No worktree selected";
-        const worktree = getCurrentViewStore().getState().worktrees.get(worktreeId);
-        if (typeof worktree?.prUrl !== "string" || worktree.prUrl.trim().length === 0)
-          return "Worktree has no associated PR";
-        return undefined;
-      },
-    })
-  );
-
-  actions.set("worktree.compareDiff", () =>
-    defineAction({
-      id: "worktree.compareDiff",
-      title: "Compare Worktrees",
-      description: "Open cross-worktree diff comparison to review changes between two branches",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ worktreeId: z.string().optional() }).optional(),
-      run: async (args, ctx: ActionContext) => {
-        const worktreeId = args?.worktreeId;
-        const targetWorktreeId =
-          worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId ?? null;
-        useWorktreeSelectionStore.getState().openCrossWorktreeDiff(targetWorktreeId);
-      },
-    })
-  );
-
-  actions.set("worktree.openIssueInPortal", () =>
-    defineAction({
-      id: "worktree.openIssueInPortal",
-      title: "Open Worktree Issue in Portal",
-      description: "Open the worktree's GitHub issue in the integrated browser",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ worktreeId: z.string().optional() }).optional(),
-      run: async (args, ctx: ActionContext) => {
-        const worktreeId = args?.worktreeId;
-        const targetWorktreeId = worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!targetWorktreeId) return;
-
-        const worktree = getCurrentViewStore().getState().worktrees.get(targetWorktreeId);
-        if (!worktree?.issueNumber) return;
-
-        const issueUrl = await githubClient.getIssueUrl(worktree.path, worktree.issueNumber);
-        if (!issueUrl) return;
-
-        await actionService.dispatch(
-          "portal.openUrl",
-          {
-            url: issueUrl,
-            title: worktree.issueTitle || `Issue #${worktree.issueNumber}`,
-            background: false,
-          },
-          { source: "user" }
-        );
-      },
-      isEnabled: (ctx: ActionContext) => {
-        const worktreeId = ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!worktreeId) return false;
-        const worktree = getCurrentViewStore().getState().worktrees.get(worktreeId);
-        return typeof worktree?.issueNumber === "number" && worktree.issueNumber > 0;
-      },
-      disabledReason: (ctx: ActionContext) => {
-        const worktreeId = ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!worktreeId) return "No worktree selected";
-        const worktree = getCurrentViewStore().getState().worktrees.get(worktreeId);
-        if (typeof worktree?.issueNumber !== "number" || worktree.issueNumber <= 0)
-          return "Worktree has no associated issue";
-        return undefined;
-      },
-    })
-  );
-
-  actions.set("worktree.resource.provision", () =>
-    defineAction({
-      id: "worktree.resource.provision",
-      title: "Provision Resource",
-      description: "Run resource provisioning commands for a worktree",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ worktreeId: z.string().optional() }).optional(),
-      isEnabled: (ctx: ActionContext) => {
-        const worktreeId = ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!worktreeId) return false;
-        const worktree = getCurrentViewStore().getState().worktrees.get(worktreeId);
-        return !!worktree?.hasProvisionCommand;
-      },
-      disabledReason: (ctx: ActionContext) => {
-        const worktreeId = ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!worktreeId) return "No worktree selected";
-        const worktree = getCurrentViewStore().getState().worktrees.get(worktreeId);
-        if (!worktree?.hasProvisionCommand) return "Worktree has no provision command configured";
-        return undefined;
-      },
-      run: async (args, ctx: ActionContext) => {
-        const worktreeId = args?.worktreeId;
-        const targetWorktreeId = worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!targetWorktreeId) throw new Error("No worktree selected");
-        try {
-          await worktreeClient.resourceAction(targetWorktreeId, "provision");
-        } catch (err) {
-          notifyWorktreeResourceError(err, "Provision failed", "Resource provisioning failed");
-        }
-      },
-    })
-  );
-
-  actions.set("worktree.resource.teardown", () =>
-    defineAction({
-      id: "worktree.resource.teardown",
-      title: "Teardown Resource",
-      description: "Run resource teardown commands for a worktree",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ worktreeId: z.string().optional() }).optional(),
-      isEnabled: (ctx: ActionContext) => {
-        const worktreeId = ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!worktreeId) return false;
-        const worktree = getCurrentViewStore().getState().worktrees.get(worktreeId);
-        return !!worktree?.hasTeardownCommand;
-      },
-      disabledReason: (ctx: ActionContext) => {
-        const worktreeId = ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!worktreeId) return "No worktree selected";
-        const worktree = getCurrentViewStore().getState().worktrees.get(worktreeId);
-        if (!worktree?.hasTeardownCommand) return "Worktree has no teardown command configured";
-        return undefined;
-      },
-      run: async (args, ctx: ActionContext) => {
-        const worktreeId = args?.worktreeId;
-        const targetWorktreeId = worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!targetWorktreeId) throw new Error("No worktree selected");
-        try {
-          await worktreeClient.resourceAction(targetWorktreeId, "teardown");
-        } catch (err) {
-          notifyWorktreeResourceError(err, "Teardown failed", "Resource teardown failed");
-        }
-      },
-    })
-  );
-
-  actions.set("worktree.resource.resume", () =>
-    defineAction({
-      id: "worktree.resource.resume",
-      title: "Resume Resource",
-      description: "Resume the resource associated with a worktree",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ worktreeId: z.string().optional() }).optional(),
-      isEnabled: (ctx: ActionContext) => {
-        const worktreeId = ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!worktreeId) return false;
-        const worktree = getCurrentViewStore().getState().worktrees.get(worktreeId);
-        return !!worktree?.hasResumeCommand;
-      },
-      disabledReason: (ctx: ActionContext) => {
-        const worktreeId = ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!worktreeId) return "No worktree selected";
-        const worktree = getCurrentViewStore().getState().worktrees.get(worktreeId);
-        if (!worktree?.hasResumeCommand) return "Worktree has no resume command configured";
-        return undefined;
-      },
-      run: async (args, ctx: ActionContext) => {
-        const worktreeId = args?.worktreeId;
-        const targetWorktreeId = worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!targetWorktreeId) throw new Error("No worktree selected");
-        try {
-          await worktreeClient.resourceAction(targetWorktreeId, "resume");
-        } catch (err) {
-          notifyWorktreeResourceError(err, "Resume failed", "Resource resume failed");
-        }
-      },
-    })
-  );
-
-  actions.set("worktree.resource.pause", () =>
-    defineAction({
-      id: "worktree.resource.pause",
-      title: "Pause Resource",
-      description: "Pause the resource associated with a worktree",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ worktreeId: z.string().optional() }).optional(),
-      isEnabled: (ctx: ActionContext) => {
-        const worktreeId = ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!worktreeId) return false;
-        const worktree = getCurrentViewStore().getState().worktrees.get(worktreeId);
-        return !!worktree?.hasPauseCommand;
-      },
-      disabledReason: (ctx: ActionContext) => {
-        const worktreeId = ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!worktreeId) return "No worktree selected";
-        const worktree = getCurrentViewStore().getState().worktrees.get(worktreeId);
-        if (!worktree?.hasPauseCommand) return "Worktree has no pause command configured";
-        return undefined;
-      },
-      run: async (args, ctx: ActionContext) => {
-        const worktreeId = args?.worktreeId;
-        const targetWorktreeId = worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!targetWorktreeId) throw new Error("No worktree selected");
-        try {
-          await worktreeClient.resourceAction(targetWorktreeId, "pause");
-        } catch (err) {
-          notifyWorktreeResourceError(err, "Pause failed", "Resource pause failed");
-        }
-      },
-    })
-  );
-
-  actions.set("worktree.resource.status", () =>
-    defineAction({
-      id: "worktree.resource.status",
-      title: "Check Resource Status",
-      description: "Check the status of the resource associated with a worktree",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ worktreeId: z.string().optional() }).optional(),
-      run: async (args, ctx: ActionContext) => {
-        const worktreeId = args?.worktreeId;
-        const targetWorktreeId = worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!targetWorktreeId) throw new Error("No worktree selected");
-        const worktree = getCurrentViewStore().getState().worktrees.get(targetWorktreeId);
-        if (!worktree) throw new Error("Worktree not found");
-        if (!worktree.hasStatusCommand) {
-          return { configured: false, status: null } as const;
-        }
-        try {
-          await worktreeClient.resourceAction(targetWorktreeId, "status");
-        } catch (err) {
-          notifyWorktreeResourceError(err, "Status check failed", "Resource status check failed");
-        }
-        const updated = getCurrentViewStore().getState().worktrees.get(targetWorktreeId);
-        return { configured: true, status: updated?.resourceStatus ?? null } as const;
-      },
-    })
-  );
-
-  actions.set("worktree.resource.connect", () =>
-    defineAction({
-      id: "worktree.resource.connect",
-      title: "Connect to Resource",
-      description: "Open a terminal session connected to the worktree's remote resource",
-      category: "worktree",
-      kind: "command",
-      danger: "safe",
-      scope: "renderer",
-      argsSchema: z.object({ worktreeId: z.string().optional() }).optional(),
-      isEnabled: (ctx: ActionContext) => {
-        const worktreeId = ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!worktreeId) return false;
-        const worktree = getCurrentViewStore().getState().worktrees.get(worktreeId);
-        return !!worktree?.resourceConnectCommand;
-      },
-      disabledReason: (ctx: ActionContext) => {
-        const worktreeId = ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!worktreeId) return "No worktree selected";
-        const worktree = getCurrentViewStore().getState().worktrees.get(worktreeId);
-        if (!worktree?.resourceConnectCommand)
-          return "Worktree has no resource connect command configured";
-        return undefined;
-      },
-      run: async (args, ctx: ActionContext) => {
-        const worktreeId = args?.worktreeId;
-        const targetWorktreeId = worktreeId ?? ctx.focusedWorktreeId ?? ctx.activeWorktreeId;
-        if (!targetWorktreeId) throw new Error("No worktree selected");
-        const worktree = getCurrentViewStore().getState().worktrees.get(targetWorktreeId);
-        if (!worktree) throw new Error("Worktree not found");
-        const connectCommand = worktree.resourceConnectCommand;
-        if (!connectCommand)
-          throw new Error("No resource connect command configured for this worktree");
-
-        await callbacks.onAddTerminal({
-          kind: "terminal",
-          cwd: worktree.path,
-          command: connectCommand,
-          title: `Connect: ${worktree.name}`,
-          location: "grid",
-          worktreeId: targetWorktreeId,
-        });
-      },
-    })
-  );
+  registerWorktreeQueryActions(actions, callbacks);
+  registerWorktreeServiceActions(actions, callbacks);
+  registerWorktreeCreateActions(actions, callbacks);
+  registerWorktreeNavigationActions(actions, callbacks);
+  registerWorktreeContextActions(actions, callbacks);
+  registerWorktreeGitHubActions(actions, callbacks);
+  registerWorktreeResourceActions(actions, callbacks);
 }
